--- poc/steps/data_validators.py
@@ -1,6 +1,0 @@
-from poc.step import step
-
-
-@step
-def data_validator(dataset_context: dict):
-    return dataset_context
