--- poc/steps/weights_preparators.py
@@ -1,6 +1,0 @@
-from poc.step import step
-
-
-@step
-def weights_preparator(weights_path: str):
-    return weights_path
