+++ src/steps/data_validators.py
@@ -1,0 +1,6 @@
+from src import step
+
+
+@step
+def data_validator(dataset_context: dict):
+    return dataset_context
