+++ src/steps/weights_preparators.py
@@ -1,0 +1,6 @@
+from src import step
+
+
+@step
+def weights_preparator(weights_path: str):
+    return weights_path
